% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{djsir_colours}
\alias{djsir_colours}
\alias{djpr_cobalt}
\alias{djpr_bondi_blue}
\alias{djpr_lima}
\alias{djpr_electric_lime}
\alias{djpr_royal_blue}
\alias{djpr_green}
\alias{djpr_persimmon}
\alias{djpr_dark_tangerine}
\alias{djpr_black_rock}
\alias{djpr_cool_grey_11}
\alias{djpr_cool_grey_1}
\title{DJSIR colours}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
djpr_cobalt

djpr_bondi_blue

djpr_lima

djpr_electric_lime

djpr_royal_blue

djpr_green

djpr_persimmon

djpr_dark_tangerine

djpr_black_rock

djpr_cool_grey_11

djpr_cool_grey_1
}
\description{
A set of colours and their # values, as defined by the DJSIR colour scheme
}
\examples{
colours()
}
\keyword{datasets}
