% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djsir_pal.R
\name{djsir_pal}
\alias{djsir_pal}
\title{Title Function takes an integer and returns colours}
\usage{
djsir_pal(n = 1)
}
\arguments{
\item{n}{numeric. The number of levels in your colour scale = number of colours required.
Minimum and default value is 1, maximum is 11. More than 6 is not recommended.
If n is >11, the colour palette is doubled, starting from 1 again.}
}
\value{
palette which is a character vector of colours
}
\description{
Title Function takes an integer and returns colours
}
\examples{
library(ggplot2)

p <- ggplot(mtcars, aes(x = wt, y = mpg, col = factor(cyl))) +
  geom_point() +
  theme_djsir() +
  scale_colour_manual(values = djsir_pal(n = 3))

p

# Alternatively, use scale_colour_djsir(), which is a wrapper
# around scale_colour_manual():

p <- ggplot(mtcars, aes(x = wt, y = mpg, col = factor(cyl))) +
  geom_point() +
  theme_djsir() +
  scale_colour_djsir(3)

p


# When applying colours as a fill, use scale_fill_djsir, which is
# a wrapper around scale_colour_manual:

p <- ggplot(mtcars, aes(x = mpg, fill = factor(cyl))) +
  geom_histogram() +
  theme_djsir() +
  scale_fill_djsir(3)
p
}
