% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djsir_pal.R
\docType{data}
\name{djsir_palette}
\alias{djsir_palette}
\title{DJSIR colour palette definition (11 colours)}
\format{
An object of class \code{character} of length 11.
}
\usage{
djsir_palette
}
\value{
character vector
}
\description{
DJSIR colour palette definition (11 colours)
}
\keyword{datasets}
