% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djsir_use_fonts.R
\name{djsir_use_fonts}
\alias{djsir_use_fonts}
\title{Enable and use the custom VIC font in plots}
\usage{
djsir_use_fonts()
}
\description{
After calling this, ggplot outputs will use the VIC font.
}
\details{
This requires the suggested packages `showtext` and `sysfonts` to work.
}
