% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djsirify_geom_defaults.R
\name{djsirify_geom_defaults}
\alias{djsirify_geom_defaults}
\title{Title Definition of defaults for individual geoms in a style guide-consistent way}
\usage{
djsirify_geom_defaults()
}
\description{
Title Definition of defaults for individual geoms in a style guide-consistent way
}
