% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djsir_pal.R
\name{regular_palette}
\alias{regular_palette}
\title{Title Defines what colours to be used depending on how many are needed}
\usage{
regular_palette(n)
}
\arguments{
\item{n}{numeric. The number of levels in your colour scale = number of colours required.
Minimum and default value is 1, maximum is 11. More than 6 is not recommended.}

\item{...}{arguments passed to `ggplot2::scale_*_manual()`}
}
\value{
palette
}
\description{
Title Defines what colours to be used depending on how many are needed
}
\examples{
TBD
}
