% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_djsir.R
\name{theme_djsir}
\alias{theme_djsir}
\title{Style your ggplot2 plots in the DJSIR style.}
\usage{
theme_djsir(
  base_size = 14,
  base_family = getOption("djsirtheme.base_font_family", "sans"),
  chart_type = "normal",
  flipped = FALSE,
  background = "white",
  panel_borders = FALSE
)
}
\arguments{
\item{base_size}{Size for text elements. Default is 14.}

\item{base_family}{Font family for text elements. Default is "sans".}

\item{chart_type}{"normal" by default. Set to "scatter" for scatter plots.}

\item{flipped}{FALSE by default. Set to TRUE if using coord_flip(). If set to
TRUE, the theme will show a vertical axis line, ticks & panel grid, while
hiding the horizontals. Ignored for type = "scatter".}

\item{background}{"white" by default.}

\item{panel_borders}{`FALSE` by default. Set to `TRUE` to enable a black
border around the plotting area.}
}
\description{
Style your ggplot2 plots in the DJSIR style.
}
