% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_djsir_base.R
\name{theme_djsir_base}
\alias{theme_djsir_base}
\title{Title Defines DJSIR base theme - function is called by 'theme_djsir' but not exported}
\usage{
theme_djsir_base(
  base_size = 12,
  base_family = "sans",
  base_line_size = points_to_mm(0.75),
  base_rect_size = points_to_mm(1),
  background = "white",
  panel_borders = FALSE
)
}
\arguments{
\item{base_size}{Default is 12}

\item{base_family}{Default is 'sans'}

\item{base_line_size}{Size of the base line}

\item{base_rect_size}{Size of the rectangle}

\item{background}{Default is 'white'}

\item{panel_borders}{Default is FALSE}
}
\value{
ret
}
\description{
Title Defines DJSIR base theme - function is called by 'theme_djsir' but not exported
}
