% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_djsir_normal.R
\name{theme_djsir_normal}
\alias{theme_djsir_normal}
\title{Title Defines DJSIR normal theme - function is called by 'theme_djsir' but not exported}
\usage{
theme_djsir_normal(
  base_size = 12,
  base_family = "sans",
  flipped = FALSE,
  background = "white",
  panel_borders = FALSE
)
}
\arguments{
\item{base_size}{Default is 12}

\item{base_family}{Default is 'sans'}

\item{flipped}{Default is FALSE}

\item{background}{Default is 'white'}

\item{panel_borders}{Default is FALSE}
}
\value{
ret
}
\description{
Title Defines DJSIR normal theme - function is called by 'theme_djsir' but not exported
}
