% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_djsir_scatter.R
\name{theme_djsir_scatter}
\alias{theme_djsir_scatter}
\title{Title Defines DJSIR theme for scatter plots - function is called by 'theme_djsir' but not exported}
\usage{
theme_djsir_scatter(
  base_size = 12,
  base_family = "sans",
  background = "white",
  panel_borders = FALSE
)
}
\arguments{
\item{base_size}{Default is 12}

\item{base_family}{Default is 'sans'}

\item{background}{Default is 'white'}

\item{panel_borders}{Default is FALSE}
}
\value{
ret
}
\description{
Title Defines DJSIR theme for scatter plots - function is called by 'theme_djsir' but not exported
}
