% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{djsir_colours}
\alias{djsir_colours}
\alias{djsir_cobalt}
\alias{djsir_bondi_blue}
\alias{djsir_lima}
\alias{djsir_electric_lime}
\alias{djsir_royal_blue}
\alias{djsir_green}
\alias{djsir_persimmon}
\alias{djsir_dark_tangerine}
\alias{djsir_black_rock}
\alias{djsir_cool_grey_11}
\alias{djsir_cool_grey_1}
\title{DJSIR colours}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
djsir_cobalt

djsir_bondi_blue

djsir_lima

djsir_electric_lime

djsir_royal_blue

djsir_green

djsir_persimmon

djsir_dark_tangerine

djsir_black_rock

djsir_cool_grey_11

djsir_cool_grey_1
}
\description{
A set of colours and their # values, as defined by the DJSIR colour scheme
}
\examples{
colours()
}
\keyword{datasets}
