% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djsir_pal.R
\name{scale_colour_djsir}
\alias{scale_colour_djsir}
\title{Title DJSIR function defining the colour of a graph object}
\usage{
scale_colour_djsir(...)
}
\arguments{
\item{...}{arguments passed to `ggplot2::scale_*_manual()`}
}
\description{
Title DJSIR function defining the colour of a graph object
}
\examples{
TBD
}
