% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djsir_pal.R
\name{scale_fill_djsir}
\alias{scale_fill_djsir}
\title{Title DJSIR wrapping function defining colour for the fill of a graph object}
\usage{
scale_fill_djsir(...)
}
\arguments{
\item{...}{arguments passed to `ggplot2::scale_*_manual()`}
}
\description{
Title DJSIR wrapping function defining colour for the fill of a graph object
}
\examples{
TBD
}
